change the directly of router, move fetchht to the /api
